backend/requirements.txt
absl-py==2.4.0
asgiref==3.11.1
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
dnspython==2.8.0
email-validator==2.1.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==25.12.19
fonttools==4.61.1
future==1.0.0
geocoder==1.38.1
greenlet==3.3.2
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.32
MouseInfo==0.1.3
numpy==2.4.2
opencage==2.4.0
opencv-contrib-python==4.13.0.90
opencv-python==4.13.0.90
packaging==26.0
phonenumbers==9.0.24
pillow==12.1.0
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pypandoc==1.16.2
pyparsing==3.3.2
pyperclip==1.11.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.2
PyQt6_sip==13.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytweening==1.2.0
pywin32==311
ratelim==0.1.6
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sounddevice==0.5.5
SQLAlchemy==2.0.48
sqlparse==0.5.5
threadpoolctl==3.6.0
tkintermapview==1.29
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.6
whitenoise==6.12.0
WTForms==3.1.1
